backend/requirements.txt

fastapi==0.112.2
pydantic==2.8.2
pydash==8.0.3
pymongo==4.8.0
pytz==2024.1
uvicorn==0.30.6
redis
pynacl
pyjwt
python-dotenv